优化

+++ trunk/sqltoy-orm/src/main/java/org/sagacity/sqltoy/dao/SqlToyLazyDao.java
@@ -607,7 +607,7 @@ public interface SqlToyLazyDao {
 
 	public <T> List<T> selectList(Class<T> resultType, EntityQuery entityQuery, DataSource dataSource);
 
-	public <T> Integer deleteByQuery(Class<T> entityClass, EntityQuery entityQuery);
+	public <T> Long deleteByQuery(Class<T> entityClass, EntityQuery entityQuery);
 
-	public <T> Integer deleteByQuery(Class<T> entityClass, EntityQuery entityQuery, DataSource dataSource);
+	public <T> Long deleteByQuery(Class<T> entityClass, EntityQuery entityQuery, DataSource dataSource);
 }
